Fixed guidelines

+++ src/helper/sodipodi-ctrlrect.h
@@ -52,7 +52,7 @@ struct _SPCtrlRectClass {
 /* Standard Gtk function */
 GtkType sp_ctrlrect_get_type (void);
 
-void sp_ctrlrect_set_area (SPCtrlRect *rect, gint x0, gint y0, gint x1, gint y1);
+void sp_ctrlrect_set_area (SPCtrlRect *rect, double x0, double y0, double x1, double y1);
 void sp_ctrlrect_set_color (SPCtrlRect *rect, guint32 border_color, gboolean has_fill, guint32 fill_color);
 void sp_ctrlrect_set_shadow (SPCtrlRect *rect, gint shadow_size, guint32 shadow_color);
